test: migrate guards keep-alive
